+++ src/main.rs
@@ -92,6 +92,8 @@ macro_rules! command_list {
     scrape        Scrape HTML into CSV data
     reverse       Reverse rows of CSV data
     transpose (t) Transpose CSV file
+    pivot         Stack multiple columns into fewer columns
+    unpivot       Split distinct values of a column into their own columns
 
 ## Split a CSV file into multiple
     split       Split CSV data into chunks
@@ -285,6 +287,7 @@ enum Command {
     P,
     Parallel,
     Partition,
+    Pivot,
     Plot,
     Progress,
     Range,
@@ -306,6 +309,7 @@ enum Command {
     Transform,
     Transpose,
     T,
+    Unpivot,
     V,
     View,
     Vocab,
@@ -367,6 +371,7 @@ impl Command {
             Command::Merge => cmd::merge::run(argv),
             Command::Parallel | Command::P => cmd::parallel::run(argv),
             Command::Partition => cmd::partition::run(argv),
+            Command::Pivot => cmd::pivot::run(argv),
             Command::Plot => cmd::plot::run(argv),
             Command::Progress => cmd::progress::run(argv),
             Command::Range => cmd::range::run(argv),
@@ -387,6 +392,7 @@ impl Command {
             Command::Top => cmd::top::run(argv),
             Command::Transform => cmd::transform::run(argv),
             Command::Transpose | Command::T => cmd::transpose::run(argv),
+            Command::Unpivot => cmd::unpivot::run(argv),
             Command::View | Command::V => cmd::view::run(argv),
             Command::Vocab => cmd::vocab::run(argv),
             Command::Window => cmd::window::run(argv),